rename: 파일명 수정, 불필요한 dummy 데이터 삭제
