+++ utils/fragranceList.ts
@@ -1,0 +1,13 @@
+import { Fragrance } from '@/types';
+
+export const fragranceList: Fragrance[] = [
+  { id: 1, name: 'Citrus' },
+  { id: 2, name: 'Fruits' },
+  { id: 3, name: 'Flowers' },
+  { id: 4, name: 'Green, Herbs' },
+  { id: 5, name: 'Spices' },
+  { id: 6, name: 'Sweet, Gourmand' },
+  { id: 7, name: 'Wood' },
+  { id: 8, name: 'Musk, Amber, Animalic' },
+  { id: 9, name: 'etc' },
+];
